+++ sila_cetoni/io/sila/io_service/feature_implementations/digitalinchannelprovider_impl.py
@@ -7,11 +7,11 @@ from queue import Queue
 from threading import Event
 from typing import Any, Dict, List, Optional, Union
 
-from qmixsdk.qmixdigio import DigitalInChannel
 from sila2.framework import Command, Feature, FullyQualifiedIdentifier, Metadata, Property
 from sila2.server import MetadataDict, SilaServer
-
 from sila_cetoni.application.system import ApplicationSystem
+
+from sila_cetoni.io.device_drivers import DigitalInChannelInterface
 
 from ..generated.digitalinchannelprovider import (
     DigitalInChannelProviderBase,
@@ -23,12 +23,12 @@ from ..generated.digitalinchannelprovider import (
 
 class DigitalInChannelProviderImpl(DigitalInChannelProviderBase):
     __system: ApplicationSystem
-    __channels: List[DigitalInChannel]
+    __channels: List[DigitalInChannelInterface]
     __channel_index_metadata: FullyQualifiedIdentifier
     __state_queues: List[Queue[State]]  # same number of items and order as `__channels`
     __stop_event: Event
 
-    def __init__(self, server: SilaServer, channels: List[DigitalInChannel], executor: Executor):
+    def __init__(self, server: SilaServer, channels: List[DigitalInChannelInterface], executor: Executor):
         super().__init__(server)
         self.__system = ApplicationSystem()
         self.__channels = channels
@@ -40,23 +40,19 @@ class DigitalInChannelProviderImpl(DigitalInChannelProviderBase):
             self.__state_queues += [Queue()]
 
             # initial value
-            self.update_State(
-                "On" if self.__channels[i].is_on() else "Off",
-                queue=self.__state_queues[i],
-            )
+            self.update_State(self.__channels[i].state, queue=self.__state_queues[i])
 
             executor.submit(self.__make_state_updater(i), self.__stop_event)
 
     def __make_state_updater(self, i: int):
         def update_state(stop_event: Event):
-            new_is_on = is_on = self.__channels[i].is_on()
-            while not stop_event.is_set():
+            new_state = state = self.__channels[i].state
+            while not stop_event.wait(0.01):
                 if self.__system.state.is_operational():
-                    new_is_on = self.__channels[i].is_on()
-                if new_is_on != is_on:
-                    is_on = new_is_on
-                    self.update_State("On" if is_on else "Off", queue=self.__state_queues[i])
-                time.sleep(0.1)
+                    new_state = self.__channels[i].state
+                if new_state != state:
+                    state = new_state
+                    self.update_State(state, queue=self.__state_queues[i])
 
         return update_state